+++ gsheet_pandas/asyncio/adapter/connection.py
@@ -1,4 +1,3 @@
-import datetime
 import json
 import logging
 import os.path
@@ -6,14 +5,14 @@ from pathlib import Path
 from typing import Literal, Optional
 
 import pandas as pd
-from pandas import Timestamp
-from pandas._libs.lib import Decimal
 
 from google.auth.transport.requests import Request
 from google.oauth2.credentials import Credentials
 from google_auth_oauthlib.flow import InstalledAppFlow
 from aiogoogle import Aiogoogle
 from aiogoogle.auth.creds import ServiceAccountCreds, UserCreds
+
+from gsheet_pandas.utils import _fix_dtypes
 
 logger = logging.getLogger("gsheet-pandas")
 
@@ -46,23 +45,6 @@ async def setup(credentials_dir: Path, token_dir: Path = None):
 
     pandas.DataFrame.to_gsheet_async = inner_generator()
     pandas.from_gsheet_async = async_drive_connection.download
-
-
-def _fix_dtypes(df: pd.DataFrame) -> pd.DataFrame:
-    """
-    Fix data types in DataFrame for proper upload to Google Sheets.
-
-    :param df: DataFrame to process
-    :return: Processed DataFrame
-    """
-    df = df.fillna("")
-    df = df.map(
-        lambda x: str(x)
-        if isinstance(x, (Timestamp, datetime.datetime, datetime.date))
-        else x
-    )
-    df = df.map(lambda x: float(x) if isinstance(x, Decimal) else x)
-    return df
 
 
 class AsyncDriveConnection:
